requirements.txt
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.1
tqdm>=4.66.0
fastapi>=0.116.1
uvicon>=0.35.0
